feat(YouTube Music): Add `Spoof video streams` patch
