+++ patches/src/main/kotlin/app/revanced/patches/music/misc/spoof/SpoofVideoStreamsPatch.kt
@@ -1,0 +1,7 @@
+package app.revanced.patches.music.misc.spoof
+
+import app.revanced.patches.shared.misc.spoof.spoofVideoStreamsPatch
+
+val spoofVideoStreamsPatch = spoofVideoStreamsPatch({
+    compatibleWith("com.google.android.apps.youtube.music")
+})
